test (Caching): Add tests that ensure `ttls` prop still works
The `ttls` property acts the same `timeTolives.byMethod`. They both take in a simple pair of string that uniquely identifies the method (kayn provides constants for it) and ttl (in milliseconds).

The key difference is that for backwards-compatibility's sake, the `ttls` prop will be the source of truth for kayn, and so it's important that it overrides all other ttl setters.