fix (DDragonRequest): Fix implicit version interaction with helper requests
Methods of the same nature as DDragon.Champion.listDataById() uses a hack within the request code where it basically intercepts the data, transforms it, and then returns that to the user.

Problem was implicit versioning calls the Realm.list() method first, and this Realm.list() data is then funneled into these data transformers that are meant for {champion|item|etc} JSONs.

This commit thus adds a flag to the request function that checks if this request is merely a helper request (which is the case of the realm pre-request, used for implicit versions).